102 tests split and working
