+++ tests/integration/test_tap_tempo.py
@@ -2,30 +2,29 @@
 
 from unittest.mock import MagicMock
 
+from tests.types import SystemFixture
 
-def test_v3_set_mod_tap_tempo(v3_system, get_urls):
+
+def test_set_mod_tap_tempo(modhandler_system: SystemFixture, get_urls):
     """set_mod_tap_tempo() POSTs the BPM value to /set_bpm."""
-    handler, _, _, _, mock_post = v3_system
+    handler, _, _, _, mock_post = modhandler_system
 
     handler.set_mod_tap_tempo(120)
 
     assert any("set_bpm" in u for u in get_urls(mock_post))
-    payload = mock_post.call_args[1].get("json") or mock_post.call_args[0][1]
-    assert payload == {"value": 120}
+    assert mock_post.call_args[1].get("json") == {"value": 120}
 
 
-def test_v3_set_mod_tap_tempo_none(v3_system, get_urls):
-    """set_mod_tap_tempo(None) is a no-op — no HTTP call."""
-    handler, _, _, _, mock_post = v3_system
-
+def test_set_mod_tap_tempo_none(modhandler_system: SystemFixture, get_urls):
+    """set_mod_tap_tempo(None) is a no-op."""
+    handler, _, _, _, mock_post = modhandler_system
     handler.set_mod_tap_tempo(None)
-
     assert not any("set_bpm" in u for u in get_urls(mock_post))
 
 
-def test_v3_get_bpm(v3_system, get_urls):
+def test_get_bpm(modhandler_system: SystemFixture, get_urls):
     """get_bpm() GETs /get_bpm and returns the parsed float."""
-    handler, _, _, mock_get, _ = v3_system
+    handler, _, _, mock_get, _ = modhandler_system
 
     def get_side_effect(url, **kwargs):
         resp = MagicMock()
@@ -35,15 +34,13 @@ def test_v3_get_bpm(v3_system, get_urls):
 
     mock_get.side_effect = get_side_effect
 
-    result = handler.get_bpm()
-
+    assert handler.get_bpm() == 120.0
     assert any("get_bpm" in u for u in get_urls(mock_get))
-    assert result == 120.0
 
 
-def test_v3_toggle_tap_tempo_enable(v3_system, snapshot):
+def test_toggle_tap_tempo_enable(modhandler_system: SystemFixture):
     """toggle_tap_tempo_enable() calls hardware and updates the LCD footswitches."""
-    handler, hw, _, mock_get, _ = v3_system
+    handler, _, _, mock_get, _ = modhandler_system
 
     def get_side_effect(url, **kwargs):
         resp = MagicMock()
@@ -52,7 +49,4 @@ def test_v3_toggle_tap_tempo_enable(v3_system, snapshot):
         return resp
 
     mock_get.side_effect = get_side_effect
-
     handler.toggle_tap_tempo_enable()
-
-    snapshot()
